requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.9.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
asyncpg>=0.30.0
alembic==1.12.1
psycopg2-binary>=2.9.10

# PostgreSQL Vector Extension
pgvector==0.2.4

# Agno Framework
agno==2.3.24

# LLM & Embeddings
openai>=2.14.0

# Vector Operations
numpy>=1.26.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6

